+++ .gitignore
@@ -1,6 +1,5 @@
 ### IntelliJ IDEA ###
 out/
-
 !**/src/main/**/out/
 !**/src/test/**/out/
 
@@ -28,11 +27,3 @@ bin/
 
 ### Mac OS ###
 .DS_Store
-
-# Maven
-log/
-target/
-
-# App
-!**/data/
-changelog.txt
