--- pom.xml
@@ -1,18 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <name>BankAccountManagement</name>
-  <url>https://github.com/fullentropppy/BankAccountManagement</url>
-
-  <groupId>ru.dgritsenko</groupId>
-  <artifactId>BankAccountManagement</artifactId>
-  <version>1.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>24</maven.compiler.source>
-    <maven.compiler.target>24</maven.compiler.target>
-  </properties>
-</project>
