+++ run/run_bam.bat
@@ -1,0 +1,9 @@
+@echo off
+setlocal
+
+set JAR_FILE=D:\IT\Repositories\BankAccountManagement\run\bam.jar
+
+echo Running the application...
+java -jar "%JAR_FILE%"
+
+endlocal
