--- src/main/java/ru/dgritsenko/userinterface/UserInterface.java
@@ -1,14 +1,0 @@
-package ru.dgritsenko.userinterface;
-
-/**
- * Интерфейс пользовательского интерфейса банковского приложения.
- * <p>Определяет основной метод для запуска взаимодействия с пользователем.
- * Реализации могут предоставлять различные способы взаимодействия
- * (консольный, графический интерфейс и т.д.).
- */
-public interface UserInterface {
-    /**
-     * Запускает пользовательский интерфейс приложения.
-     */
-    void run();
-}
