+++ src/ru/dgritsenko/bam/bank/TransactionStatus.java
@@ -1,0 +1,65 @@
+package ru.dgritsenko.bam.bank;
+
+/**
+ * Перечисление, представляющее возможные статусы транзакции.
+ */
+public enum TransactionStatus {
+    UNCOMMITTED("Не подтверждена"),
+    COMMITTED("Подтверждена"),
+    CANCELED("Отменена");
+
+    private final String title;
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // OVERRIDDEN
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Возвращает строковое представление статуса транзакции.
+     *
+     * @return строковое представление (название) статуса транзакции
+     */
+    @Override
+    public String toString() {
+        return getTitle();
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // CONSTRUCTORS
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Конструктор перечисления.
+     *
+     * @param title название статуса транзакции
+     */
+    TransactionStatus(String title) {
+        this.title = title;
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // GETTERS
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Возвращает название статуса транзакции.
+     *
+     * @return название статуса транзакции
+     */
+    public String getTitle() {
+        return title;
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // METHODS. GETTING DATA
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Проверяет, является ли транзакция подтвержденной.
+     *
+     * @return {@code true}, если транзакция подтверждена
+     */
+    public boolean isCommitted() {
+        return this == COMMITTED;
+    }
+}
