+++ src/ru/dgritsenko/bam/bank/TransactionType.java
@@ -1,0 +1,75 @@
+package ru.dgritsenko.bam.bank;
+
+/**
+ * Перечисление, представляющее возможные типы транзакций.
+ */
+public enum TransactionType {
+    DEPOSIT("Пополнение"),
+    CREDIT("Поступление"),
+    WITHDRAW("Снятие наличных"),
+    TRANSFER("Перевод");
+
+    private final String title;
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // OVERRIDDEN
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Возвращает строковое представление типа транзакции.
+     *
+     * @return строковое представление (название) типа транзакции
+     */
+    @Override
+    public String toString() {
+        return getTitle();
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // CONSTRUCTORS
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Конструктор перечисления.
+     *
+     * @param title название типа транзакции
+     */
+    TransactionType(String title) {
+        this.title = title;
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // GETTERS
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Возвращает название типа транзакции.
+     *
+     * @return название типа транзакции
+     */
+    public String getTitle() {
+        return title;
+    }
+
+    // -----------------------------------------------------------------------------------------------------------------
+    // METHODS. GETTING DATA
+    // -----------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Проверяет, требует ли тип транзакции указания счета получателя.
+     *
+     * @return {@code true}, если требуется счет получателя
+     */
+    public boolean hasToAccount() {
+        return this == TRANSFER;
+    }
+
+    /**
+     * Проверяет, является ли операция пополнением (увеличивает баланс).
+     *
+     * @return {@code true}, если операция увеличивает баланс
+     */
+    public boolean isAddition() {
+        return this == DEPOSIT || this == CREDIT;
+    }
+}
